+++ PWM.h
@@ -1,7 +1,6 @@
 /*!\file PWM.h
 ** \author SMFSW
-** \date 2017
-** \copyright MIT (c) 2017, SMFSW
+** \copyright MIT (c) 2017-2018, SMFSW
 ** \brief Straightforward PWM handling
 ** \warning Shall work for all STM32 F families, L families not totally covered
 **/