+++ RTC_ex.c
@@ -1,7 +1,6 @@
 /*!\file RTC_ex.c
 ** \author SMFSW
-** \date 2017
-** \copyright MIT (c) 2017, SMFSW
+** \copyright MIT (c) 2017-2018, SMFSW
 ** \brief Basic RTC handling
 **/
 /****************************************************************/