+++ pattern2d.c
@@ -1,7 +1,6 @@
 /*!\file pattern2d.c
 ** \author SMFSW
-** \date 2017
-** \copyright MIT (c) 2017, SMFSW
+** \copyright MIT (c) 2017-2018, SMFSW
 ** \brief 2 dimensional patterns utilities
 **/
 /****************************************************************/