+++ time_utils.h
@@ -1,7 +1,6 @@
 /*!\file time_utils.h
 ** \author SMFSW
-** \date 2017
-** \copyright MIT (c) 2017, SMFSW
+** \copyright MIT (c) 2017-2018, SMFSW
 ** \brief Time related utilities
 **/
 /****************************************************************/